refactor(toolkit): promote the shared toolkit to its own package
The read-only tool functions move from interloper_db.toolkit into a new
interloper-toolkit workspace package. The persistence layer stops carrying
LLM-tool-shaped facades, the toolkit gets its own release/versioning
surface, and consumers (interloper-agent, interloper-mcp) now declare the
dependency explicitly. No behavior change: contents moved verbatim, only
import paths and packaging metadata changed.

By Digitl
