+++ packages/interloper-agent/src/interloper_agent/tools/analytics.py
@@ -1,7 +1,7 @@
 """Analytics tools — thin ADK wrappers over the shared read-only toolkit.
 
 The implementations (and the LLM-facing docstrings, adopted below) live in
-``interloper_db.toolkit.analytics`` so the MCP server exposes the same logic.
+``interloper_toolkit.analytics`` so the MCP server exposes the same logic.
 """
 
 from __future__ import annotations
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import Any
 
 from google.adk.tools.tool_context import ToolContext
-from interloper_db.toolkit import analytics as toolkit_analytics
+from interloper_toolkit import analytics as toolkit_analytics
 
 from interloper_agent.context import toolkit_ctx
