+++ pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
     "interloper-mcp",
     "interloper-pandas",
     "interloper-scheduler",
+    "interloper-toolkit",
 ]
 
 [dependency-groups]
@@ -52,6 +53,7 @@ interloper-google-cloud = { workspace = true }
 interloper-agent = { workspace = true }
 interloper-app = { workspace = true }
 interloper-mcp = { workspace = true }
+interloper-toolkit = { workspace = true }
 
 [tool.uv.workspace]
 members = ["packages/*"]
@@ -170,6 +172,7 @@ version_toml = [
     "packages/interloper-agent/pyproject.toml:project.version",
     "packages/interloper-app/pyproject.toml:project.version",
     "packages/interloper-mcp/pyproject.toml:project.version",
+    "packages/interloper-toolkit/pyproject.toml:project.version",
 ]
 # Keep the Helm chart in lockstep with the Python version. `version` (chart
 # SemVer) and `appVersion` (the image tag the chart deploys) both track the
